+++ src/class/context.rs
@@ -10,6 +10,10 @@ use crate::{PyClass, PyObject};
 /// Context manager interface
 #[allow(unused_variables)]
 pub trait PyContextProtocol<'p>: PyClass {
+    #[deprecated(
+        since = "0.14.0",
+        note = "prefer implementing `__enter__` in `#[pymethods]` instead of in a protocol"
+    )]
     fn __enter__(&'p mut self) -> Self::Result
     where
         Self: PyContextEnterProtocol<'p>,
@@ -17,6 +21,10 @@ pub trait PyContextProtocol<'p>: PyClass {
         unimplemented!()
     }
 
+    #[deprecated(
+        since = "0.14.0",
+        note = "prefer implementing `__exit__` in `#[pymethods]` instead of in a protocol"
+    )]
     fn __exit__(
         &'p mut self,
         exc_type: Option<Self::ExcType>,